Fold executable launch into App
