+++ src/windows_mcp/tools/__init__.py
@@ -6,7 +6,6 @@ from windows_mcp.tools import (
     display,
     filesystem,
     input,
-    launch,
     multi,
     notification,
     process,
@@ -23,7 +22,6 @@ _MODULES = [
     filesystem,
     snapshot,
     input,
-    launch,
     scrape,
     multi,
     clipboard,
